Se agregó detalleOrden.

+++ src/main/java/com/backend/LosPostes/data/model/entity/Producto.java
@@ -30,7 +30,7 @@ public class Producto {
     private String descripcion;
 
     @Column(name = "precio") 
-    private float precio;
+    private Double precio;
 
     @Column(name = "subcategorias_subcategoriaid") 
     private Integer subcategoriaID;
